move accountDate to utils
